* inttypes-format-ex: (24 commits)
  Tweak a ULL literal added in EX
  gasnet_mmap.c: Use PRIuPTR to output uintptr_t
  template: Use PRIuPTR to output uintptr_t segsize
  Adjust static checking of TEST_SEGSZ in test.h
  Add %lf to the list of non-portable printf format specifiers
  pami: use PRI* macros for formatting 64-bit types
  ibv: use PRI* macros for formatting 64-bit types
  gemini: use PRI* macros for formatting 64-bit types
  top-level: Use PRIuPTR and uintptr_t for formatting size_t
  other: Use PRIuPTR and uintptr_t for formatting size_t
  tests: Use PRIuPTR and uintptr_t for formatting size_t
  fix AMMPI to use 64-bit stats output
  Fix overflow in timer calibration debug output
  Remove %llu and long long from conduit-independent code
  Remove %llu and long long from ibv-conduit
  Use int32_t values in testlegacy memset test
  Remove %ll[ud] and long long from GASNet tests
  Remove %llu and long long from AMMPI
  Remove %llu and long long from AMUDP
  Add (U)LL suffixed 64-bit integer literals to required C99 subset
  ...

+++ gemini-conduit/gasnet_core.c
@@ -541,7 +541,7 @@ extern uintptr_t gasnetc_MaxPinMem(uintptr_t msgspace)
   if (limit < granularity) {
     gasnetc_GNIT_Abort("Unable to alloc and pin minimal memory of size %d bytes",(int)granularity);
   }
-  GASNETI_TRACE_PRINTF(C,("MaxPinMem = %lu",(unsigned long)limit));
+  GASNETI_TRACE_PRINTF(C,("MaxPinMem = %"PRIuPTR,limit));
   return (uintptr_t)limit;
 }
 
@@ -765,8 +765,8 @@ extern int gasnetc_attach( gasnetex_client_t      *client_p,
                            int                    numentries,
                            uintptr_t              segsize)
 {
-  GASNETI_TRACE_PRINTF(C,("gasnetc_attach(table (%i entries), segsize=%lu)",
-                          numentries, (unsigned long)segsize));
+  GASNETI_TRACE_PRINTF(C,("gasnetc_attach(table (%i entries), segsize=%"PRIuPTR")",
+                          numentries, segsize));
 
   if (!gasneti_init_done) 
     GASNETI_RETURN_ERRR(NOT_INIT, "GASNet attach called before init");
