+++ tests/testlogGP.c
@@ -359,7 +359,7 @@ int main(int argc, char **argv)
     GASNET_Safe(gasnetex_ClientInit(&myclient, &myep, &myteam, &argc, &argv, "testlogGP", 0));
     GASNET_Safe(gasnetex_TeamSegmentCreate(&mysegment, myteam, NULL, TEST_SEGSZ_REQUEST, GASNETEX_MEMKIND_DEFAULT, 0));
     snprintf(usagestr, sizeof(usagestr), "iters pollcnt sizes...\n"
-                      "    sizes are limited to %ld", (long)TEST_SEGSZ);
+                      "    sizes are limited to %"PRIuPTR, (uintptr_t)TEST_SEGSZ);
     test_init("testlogGP",1, usagestr);
     
 
@@ -394,7 +394,7 @@ int main(int argc, char **argv)
         int size = atoi(argv[i]);
 
         if (size < 0 || size > TEST_SEGSZ) {
-            MSG0("size is limited to <= %ld\n", (long)TEST_SEGSZ);
+            MSG0("size is limited to <= %"PRIuPTR"\n", (uintptr_t)TEST_SEGSZ);
             continue;
         }
